+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 - refactoring of the fixture function compilation (olegpidsadnyi)
 - related factory fix (olegpidsadnyi)
+- post_generation fixture dependency fixed (olegpidsadnyi)
+- model fixture registration with specific name (olegpidsadnyi)
+- README updated (olegpidsadnyi)
 
 1.0.1
 -----